Don't pass "SkipEmptyParts" when splitting polylines
The enum value "QString::SkipEmptyParts" produces a deprecation warning
when compiling with Qt 5.15. And since actually it's not even needed
to skip empty parts of the string, let's simply remove it. It just makes
the parser slightly more strict, since now polylines containing empty
vertices or a trailing semicolon are now rejected.